Fixed file downloading, refactored auth middleware

+++ api/server/handlers.go
@@ -1,21 +1,20 @@
 package server
 
 import (
-	"MicroFileServer/logging"
 	"MicroFileServer/models"
 	"bytes"
 	"context"
 	"encoding/json"
-	"fmt"
 	"github.com/gorilla/mux"
 	log "github.com/sirupsen/logrus"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo/gridfs"
 	"go.mongodb.org/mongo-driver/mongo/options"
+	"io"
 	"io/ioutil"
 	"net/http"
-	"os"
+	"strconv"
 	"time"
 )
 
@@ -23,20 +22,8 @@ import (
 
 func downloadFile(w http.ResponseWriter, r *http.Request) {
 	var downloadedFile models.File
-	var buf bytes.Buffer
 
-	subClaim, err := getClaim(r, "sub")
-	if err != nil {
-		logging.AuthError(w, err, "getClaim(sub)")
-		return
-	}
 	data := mux.Vars(r)
-
-	if subClaim != data["id"] {
-		w.WriteHeader(403)
-		return
-	}
-
 	objID, err := primitive.ObjectIDFromHex(string(data["id"]))
 	if err != nil {
 		http.NotFound(w, r)
@@ -50,11 +37,11 @@ func downloadFile(w http.ResponseWriter, r *http.Request) {
 		http.NotFound(w, r)
 		return
 	}
-	fileName := downloadedFile.FileName
-	w.Header().Set("Content-Disposition", "attachment; filename=\""+fileName+"\"")
-	bucket, _ := gridfs.NewBucket(db)
 
-	dStream, err := bucket.DownloadToStream(objID, &buf)
+	fileName := downloadedFile.FileName
+	buf := bytes.NewBuffer(nil)
+	bucket, _ := gridfs.NewBucket(db)
+	dStream, err := bucket.DownloadToStream(objID, buf)
 	if err != nil {
 		log.WithFields(log.Fields{
 			"function" : "bucket.DownloadToStream",
@@ -63,11 +50,15 @@ func downloadFile(w http.ResponseWriter, r *http.Request) {
 		},
 		).Fatal("DB interaction resulted in error, shutting down...")
 	}
-	fmt.Printf("File size to download: %v\n", dStream)
-	ioutil.WriteFile(fileName, buf.Bytes(), 0600)
+	log.WithFields(log.Fields{
+		"fileSize" : dStream,
+	},
+	).Info("File size to download: ")
+
+	w.Header().Set("Content-Disposition", "attachment; filename=\""+fileName+"\"")
+	w.Header().Set("Content-Length", strconv.FormatInt(downloadedFile.Length, 10))
+	io.Copy(w, buf)
 }
-
-
 
 func uploadFile(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
@@ -76,59 +67,65 @@ func uploadFile(w http.ResponseWriter, r *http.Request) {
 	data, handler, err := r.FormFile("uploadingForm")
 	if err != nil {
 		w.Write([]byte("File is too big! (Max size: 30MB)"))
+		log.WithFields(log.Fields{
+			"err" : err,
+		},
+		).Info("File is too big! (Max size: 30MB)")
 		return
 	}
-
 	defer data.Close()
 	fileBytes, err := ioutil.ReadAll(data)
 	if err != nil {
-		log.Fatal(err)
-	}
-	subClaim, err := getClaim(r, "sub")
-	if err != nil {
-		logging.AuthError(w, err, "getClaim (sub)")
+		log.WithFields(log.Fields{
+			"function" : "ioutil.ReadAll(data)",
+			"handler" : "uploadFile",
+			"error"	:	err,
+		},
+		).Warn("Can't read file data!")
 		return
 	}
-
 	desc := r.FormValue("fileDescription")
 
 	gridFSOptions := options.GridFSUpload()
-	gridFSOptions.SetMetadata(bson.M{"fileSender" : subClaim, "fileDescription" : desc})
+	gridFSOptions.SetMetadata(bson.M{"fileSender" : Claims.Sub, "fileDescription" : desc})
 	bucket, err := gridfs.NewBucket(db)
 	if err != nil {
-		log.Fatal(err)
+		log.WithFields(log.Fields{
+			"function" : "gridfs.NewBucket(db)",
+			"handler" : "uploadFile",
+			"error"	:	err,
+		},
+		).Warn("Can't create new bucket!")
+		return
 	}
 	uploadStream, err := bucket.OpenUploadStream(handler.Filename, gridFSOptions)
 	if err != nil {
-		fmt.Println(err)
-		os.Exit(1)
+		log.WithFields(log.Fields{
+			"function" : "OpenUploadStream(handler.Filename, gridFSOptions)",
+			"handler" : "uploadFile",
+			"error"	:	err,
+		},
+		).Warn("Can't open upload stream!")
+		return
 	}
 	defer uploadStream.Close()
 
 	fileSize, err := uploadStream.Write(fileBytes)
 	if err != nil {
-		log.Fatal(err)
+		log.WithFields(log.Fields{
+			"function" : "uploadStream.Write(fileBytes)",
+			"handler" : "uploadFile",
+			"error"	:	err,
+		},
+		).Warn("Can't write to upload stream!")
+		return
 	}
-	log.Infof("Write file to DB was successful. File size: %dM\n", fileSize)
+	log.WithFields(log.Fields{
+		"fileSize" : fileSize,
+	},
+	).Info("Write file to DB was successful. File size: ")
 	w.Write([]byte("Successfully uploaded file!"))
-
-	/*objID, err := primitive.ObjectIDFromHex(fmt.Sprintf("%v", uploadStream.FileID))
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-	filter := bson.M{"_id" : objID}
-	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-	err = collection.FindOne(ctx, filter).Decode(&file)
-	if err != nil {
-		http.NotFound(w, r)
-		return
-	}
-
-	json.NewEncoder(w).Encode(file)*/
 }
-
-
 
 func deleteFile(w http.ResponseWriter, r *http.Request) {
 	var requiredFile models.File
@@ -147,22 +144,33 @@ func deleteFile(w http.ResponseWriter, r *http.Request) {
 		http.NotFound(w, r)
 		return
 	}
-	bucket, err := gridfs.NewBucket(db)
-	if err != nil {
-		log.Fatal(err)
-	}
-	err = bucket.Delete(objID)
+	if requiredFile.Metadata.FileSender == Claims.Sub || isAdmin() {
+		bucket, err := gridfs.NewBucket(db)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"function" : "gridfs.NewBucket(db)",
+				"handler" : "deleteFile",
+				"error"	:	err,
+			},
+			).Warn("Can't create new bucket!")
+			return
+		}
+		err = bucket.Delete(objID)
 
-	if err != nil {
-		log.WithFields(log.Fields{
-			"function" : "bucket.Delete",
-			"handler" : "deleteFile",
-			"error"	:	err,
-		},
-		).Fatal("DB interaction resulted in error, shutting down...")
+		if err != nil {
+			log.WithFields(log.Fields{
+				"function" : "bucket.Delete",
+				"handler" : "deleteFile",
+				"error"	:	err,
+			},
+			).Fatal("DB interaction resulted in error, shutting down...")
+		}
+		w.WriteHeader(200)
+		w.Write([]byte("Successfully deleted file!"))
+	} else {
+		w.WriteHeader(403)
+		return
 	}
-	w.WriteHeader(200)
-	w.Write([]byte("Successfully deleted file!"))
 }
 
 func getFilesList(w http.ResponseWriter, r *http.Request) {
@@ -196,44 +204,50 @@ func getFilesList(w http.ResponseWriter, r *http.Request) {
 
 func getFilesListForUser(w http.ResponseWriter, r *http.Request) {
 	files := make([]models.File, 0)
+
 	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+
 	data := mux.Vars(r)
 	user := data["user"]
-
-	filter := bson.M{
-		"metadata.fileSender" : user,
+	if user == Claims.Sub || isAdmin() {
+		filter := bson.M{
+			"metadata.fileSender" : user,
+		}
+		ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
+		cur, err := collection.Find(ctx, filter)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"function" : "mongo.Find",
+				"handler" : "getFilesList",
+				"error"	:	err,
+			},
+			).Fatal("DB interaction resulted in error, shutting down...")
+		}
+		ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
+		defer cur.Close(ctx)
+		ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
+		err = cur.All(ctx, &files)
+		if err != nil {
+			log.WithFields(log.Fields{
+				"function" : "mongo.All",
+				"handler" : "getFilesList",
+				"error"	:	err,
+			},
+			).Fatal("DB interaction resulted in error, shutting down...")
+		}
+		json.NewEncoder(w).Encode(files)
+	} else {
+		w.WriteHeader(403)
+		return
 	}
-
-	ctx, _ := context.WithTimeout(context.Background(), 10*time.Second)
-	cur, err := collection.Find(ctx, filter)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"function" : "mongo.Find",
-			"handler" : "getFilesList",
-			"error"	:	err,
-		},
-		).Fatal("DB interaction resulted in error, shutting down...")
-	}
-	ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
-	defer cur.Close(ctx)
-	ctx, _ = context.WithTimeout(context.Background(), 10*time.Second)
-	err = cur.All(ctx, &files)
-	if err != nil {
-		log.WithFields(log.Fields{
-			"function" : "mongo.All",
-			"handler" : "getFilesList",
-			"error"	:	err,
-		},
-		).Fatal("DB interaction resulted in error, shutting down...")
-	}
-	json.NewEncoder(w).Encode(files)
 }
 
 func getFile(w http.ResponseWriter, r *http.Request) {
 	var file models.File
-	w.Header().Set("Content-Type", "application/json; charset=utf-8")
-	data := mux.Vars(r)
 
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+
+	data := mux.Vars(r)
 	objID, err := primitive.ObjectIDFromHex(string(data["id"]))
 	if err != nil {
 		http.NotFound(w, r)
@@ -249,5 +263,10 @@ func getFile(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	json.NewEncoder(w).Encode(file)
+	if file.Metadata.FileSender == Claims.Sub || isAdmin() {
+		json.NewEncoder(w).Encode(file)
+	} else {
+		w.WriteHeader(403)
+		return
+	}
 }
